+++ src/components/MostPopular.js
@@ -1,7 +1,7 @@
 import { useState, useEffect } from 'react'
 import { Card, Button } from 'react-bootstrap'
 import axios from 'axios'
-
+import SharingButtons from './Sharing/SharingButtons'
 const nytMostPopularUrl = `https://api.nytimes.com/svc/mostpopular/v2/viewed/1.json?api-key=${process.env.REACT_APP_NYT_API_KEY}`
 
 export default function MostPopular() {
@@ -34,6 +34,9 @@ export default function MostPopular() {
               </a>
             </Button>
           </Card.Body>
+          <Card.Footer>
+            <SharingButtons url={post.url} />
+          </Card.Footer>
         </Card>
       ))}
     </div>